# docs/swagger.yaml
basePath: /
definitions:
...
  requestmodels.Incident:
    properties:
      Description:
        type: string
      Landlord:
        type: string
      Status:
        type: integer
      Tenant:
        type: string
      Type:
        type: integer
    type: object
  requestmodels.Invite:
    properties:
      Landlord:
        type: string
      Property:
        type: string
      Tenant:
        type: string
    type: object
  requestmodels.MarketProfile:
    properties:
      AvailabilityDate:
        type: string
      Deposit:
        type: number
      Description:
        type: string
      HeatPrice:
        type: number
      MinimumIncome:
        type: number
      MinimumPeriod:
        type: integer
      Period:
        type: integer
      PowerPrice:
        type: number
      Rent:
        type: number
    type: object
...
  responsemodels.Incident:
    properties:
      Description:
        type: string
      Id:
        type: string
      Landlord:
        type: string
      Property:
        type: string
      Status:
        type: integer
      Tenant:
        type: string
      Type:
        type: integer
    type: object
  responsemodels.MarketProfile:
    properties:
      AvailabilityDate:
        type: string
      Deposit:
        type: number
      Description:
        type: string
      HeatPrice:
        type: number
      Id:
        type: string
      MinimumIncome:
        type: number
      MinimumPeriod:
        type: integer
      Period:
        type: integer
      PowerPrice:
        type: number
      Rent:
        type: number
    type: object
  responsemodels.Property:
    properties:
      Apartment:
        type: string
      Balcony:
        type: boolean
      City:
        type: string
      Country:
        type: string
      Garage:
        type: boolean
      Garden:
        type: boolean
      HeatType:
        type: integer
      Housenumber:
        type: string
      Id:
        type: string
      Landlord:
        type: string
      Rooms:
        type: integer
      Size:
        type: integer
      Status:
        type: integer
      Street:
        type: string
      Tenant:
        type: string
      Type:
        type: integer
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Swagger Example API
  version: "2.0"
paths:
...
  /incident/{propertyId}:
    get:
      consumes:
      - application/json
      description: Get all incidents
      parameters:
      - description: Property ID
        in: path
        name: propertyId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
          schema:
            items:
              $ref: '#/definitions/responsemodels.Incident'
            type: array
      summary: Get all incidents
      tags:
      - incidents
  /incident/{propertyId}/{incidentId}:
    delete:
      consumes:
      - application/json
      description: Delete an incident
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Delete an incident
      tags:
      - incidents
    get:
      consumes:
      - application/json
      description: Get an incident
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
          schema:
            $ref: '#/definitions/responsemodels.Incident'
      summary: Get an incident
      tags:
      - incidents
    put:
      consumes:
      - application/json
      description: Update an incident
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      - description: Incident
        in: body
        name: incident
        required: true
        schema:
          $ref: '#/definitions/requestmodels.Incident'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Update an incident
      tags:
      - incidents
  /incidents/{propertyId}:
    post:
      consumes:
      - application/json
      description: Create a new incident
      parameters:
      - description: Incident
        in: body
        name: incident
        required: true
        schema:
          $ref: '#/definitions/requestmodels.Incident'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create a new incident
      tags:
      - incidents
...
  /property/invite/{id}/{userId}:
    post:
      consumes:
      - application/json
      description: Accept an invite
      parameters:
      - description: Invite ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Property'
      summary: Accept an invite
      tags:
      - Property
  /property/invite/{propertyId}:
    get:
      consumes:
      - application/json
      description: Create an invite
      parameters:
      - description: Invite
        in: body
        name: invite
        required: true
        schema:
          $ref: '#/definitions/requestmodels.Invite'
      produces:
      - application/json
      responses:
        "200":
          description: InviteId
          schema:
            type: string
      summary: Create an invite
      tags:
      - Property
...
  /property/marketprofile/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a market profile
      parameters:
      - description: Market Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete a market profile
      tags:
      - MarketProfile
    get:
      consumes:
      - application/json
      description: Get a market profile
      parameters:
      - description: Market Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
          schema:
            $ref: '#/definitions/responsemodels.MarketProfile'
      summary: Get a market profile
      tags:
      - MarketProfile
    post:
      consumes:
      - application/json
      description: Create a new market profile
      parameters:
      - description: Market Profile
        in: body
        name: marketProfile
        required: true
        schema:
          $ref: '#/definitions/requestmodels.MarketProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.MarketProfile'
      summary: Create a new market profile
      tags:
      - MarketProfile
    put:
      consumes:
      - application/json
      description: Update a market profile
      parameters:
      - description: Market Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: Market Profile
        in: body
        name: marketProfile
        required: true
        schema:
          $ref: '#/definitions/requestmodels.MarketProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.MarketProfile'
      summary: Update a market profile
      tags:
      - MarketProfile
...
schemes:
- http
swagger: "2.0"